search_mcp/requirements.txt
mcp>=1.3.0
anthropic>=0.18.1
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-api-python-client>=2.100.0 